refactor(login): create LoginForm component and update Login page to use it
